cleanup + minor refactoring
