+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/index.js":"1","/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/components/XMLForm.js":"2","/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/App.js":"3"},{"size":500,"mtime":1607886871533,"results":"4","hashOfConfig":"5"},{"size":2611,"mtime":1607896686770,"results":"6","hashOfConfig":"5"},{"size":278,"mtime":1607887725496,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1o63hev",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/index.js",[],["15","16"],"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/components/XMLForm.js",[],"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/App.js",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/index.js":"1","/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/components/XMLForm.js":"2","/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/App.js":"3"},{"size":500,"mtime":1607886871533,"results":"4","hashOfConfig":"5"},{"size":2494,"mtime":1607899903356,"results":"6","hashOfConfig":"5"},{"size":278,"mtime":1607887725496,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1o63hev",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/index.js",[],["15","16"],"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/components/XMLForm.js",[],"/Users/kennethfaria/Documents/DataPassports/Challenge/datapassports-challenge/src/App.js",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
